medical_insurance_charges_exploratory_data_analysis
EDA: Exploratory Data Analysis

This notebook contains the exploratory data analysis (EDA) for the Medical Cost Analysis project. The goal of this phase was to deeply understand the dataset and uncover the key factors driving medical costs.

Purpose & Skills Demonstrated
Data Validation and Cleaning: The notebook begins with data quality checks to ensure the dataset is complete and ready for analysis.

Feature-Based Insights: Using visualizations and statistical methods, the analysis focuses on the impact of specific features like smoker, age, bmi, and region on total_charges.

Hypothesis Testing: The analysis uses a statistical t-test to rigorously confirm that the difference in medical costs between smokers and non-smokers is not due to random chance.

Data Storytelling: The visualizations and narrative within the notebook tell a compelling story about the data, culminating in the clear conclusion that smoking status is the most significant factor influencing costs.

This EDA notebook is a key component of my end-to-end data project and serves as a strong example of my ability to generate actionable business intelligence from raw data.